requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-asgi-testclient==1.4.11
bcrypt==4.1.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.1
Deprecated==1.2.14
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
Faker==23.2.1
fastapi==0.109.0
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
humanize==4.9.0
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
limits==3.9.0
Mako==1.3.2
MarkupSafe==2.1.3
multidict==6.0.5
numpy==1.26.4
optional==0.0.1
orjson==3.9.12
packaging==23.2
pandas==2.2.0
passlib==1.7.4
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
pytest==8.0.0
pytest-asyncio==0.23.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
slowapi==0.1.9
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.26.0
vine==5.1.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wheel==0.40.0
wrapt==1.16.0
